requirements.txt
arrow==0.12.1
Babel==2.6.0
binaryornot==0.4.4
cached-property==1.4.3
certifi==2018.4.16
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
diff-match-patch==20121119
Django==2.1
django-admin-interface==0.6.3
django-bootstrap4==0.0.6
django-colorfield==0.1.15
django-dia==0.4.1
django-extensions==2.1.0
django-filter==2.0.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==1.0.1
django-phonenumber-field==2.0.0
django-tables2==1.21.2
docker==3.4.1
docker-compose==1.22.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.0.1
future==0.16.0
gunicorn==19.9.0
idna==2.6
jdcal==1.4
Jinja2==2.10
jinja2-time==0.2.0
jsonschema==2.6.0
MarkupSafe==1.0
odfpy==1.3.6
openpyxl==2.5.5
phonenumberslite==8.9.10
poyo==0.4.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.18.4
six==1.11.0
tablib==0.12.1
texttable==0.9.1
unicodecsv==0.14.1
urllib3==1.22
websocket-client==0.48.0
whichcraft==0.4.1
xlrd==1.1.0
xlwt==1.3.0
